# backend/src/api-integrations/api-integrations.dto.ts
import {
  IsIn,
  IsObject,
  IsOptional,
  IsString,
  IsUrl,
  MaxLength,
  MinLength,
} from 'class-validator';

const HTTP_METHODS = ['GET', 'POST', 'PUT', 'PATCH', 'DELETE'] as const;
const AUTH_TYPES = ['none', 'bearer', 'basic', 'apiKey'] as const;

export class CreateApiChannelDto {
  @IsString()
  @MinLength(2)
  @MaxLength(80)
  name!: string;

  @IsOptional()
  @IsString()
  @MaxLength(240)
  description?: string;
}

export class UpdateApiChannelDto {
  @IsOptional()
  @IsString()
  @MinLength(2)
  @MaxLength(80)
  name?: string;

  @IsOptional()
  @IsString()
  @MaxLength(240)
  description?: string;
}

export class SaveApiRequestDto {
  @IsString()
  @MinLength(2)
  @MaxLength(100)
  name!: string;

  @IsOptional()
  @IsString()
  @MaxLength(240)
  description?: string;

  @IsIn(HTTP_METHODS)
  method!: string;

  @IsUrl({ require_tld: false, protocols: ['http', 'https'] })
  @MaxLength(2000)
  url!: string;

  @IsIn(AUTH_TYPES)
  authType!: string;

  @IsOptional()
  @IsObject()
  authConfig?: Record<string, string>;

  @IsOptional()
  @IsString()
  @MaxLength(4000)
  queryParams?: string;

  @IsOptional()
  @IsObject()
  headers?: Record<string, string>;

  @IsOptional()
  @IsObject()
  variables?: Record<string, string>;

  @IsOptional()
  @IsString()
  @MaxLength(20000)
  body?: string;
}

export class ImportInsomniaDto {
  @IsString()
  @MinLength(1)
  @MaxLength(5_000_000)
  content!: string;
}