# docs/insomnia/openapi.yaml
openapi: 3.0.3
info:
  title: Assignment Tickets NAPP API
  version: 1.0.0
  description: |
    Coleção completa das rotas HTTP do backend. Importe este arquivo no Insomnia.
    Execute `Auth / Login` primeiro; o Insomnia armazenará automaticamente o cookie
    de sessão `connect.sid` usado nas rotas protegidas.
servers:
  - url: http://localhost:3001
    description: Backend local
...
paths:
...
  /api-integrations:
    get:
      tags: [API Integrations]
      summary: Listar canais e requisições salvas
      operationId: listApiIntegrations
      security: [{ cookieSession: [] }]
      responses:
        '200': { description: Canais e requisições }
  /api-integrations/import/insomnia:
    post:
      tags: [API Integrations]
      summary: Importar coleção YAML do Insomnia v5
      operationId: importInsomniaCollection
      security: [{ cookieSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, description: Conteúdo integral do arquivo YAML }
      responses:
        '201': { description: Coleção importada }
        '400': { description: Arquivo inválido ou não suportado }
  /api-integrations/channels:
    post:
      tags: [API Integrations]
      summary: Criar canal de API
      operationId: createApiChannel
      security: [{ cookieSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateApiChannel' }
            example: { name: Pedidos, description: APIs do serviço de pedidos }
      responses:
        '201': { description: Canal criado }
  /api-integrations/channels/{channelId}:
    parameters: [{ $ref: '#/components/parameters/ChannelId' }]
    patch:
      tags: [API Integrations]
      summary: Atualizar canal de API
      operationId: updateApiChannel
      security: [{ cookieSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateApiChannel' }
            example: { name: Pedidos v2, description: APIs atualizadas }
      responses:
        '200': { description: Canal atualizado }
    delete:
      tags: [API Integrations]
      summary: Excluir canal de API
      operationId: deleteApiChannel
      security: [{ cookieSession: [] }]
      responses:
        '200': { description: Canal excluído }
  /api-integrations/channels/{channelId}/requests:
    post:
      tags: [API Integrations]
      summary: Salvar requisição em um canal
      operationId: createApiRequest
      security: [{ cookieSession: [] }]
      parameters: [{ $ref: '#/components/parameters/ChannelId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaveApiRequest' }
            example:
              name: Consultar pedido
              description: Busca um pedido pelo ID
              method: GET
              url: http://localhost:8080/orders/{{orderId}}
              authType: bearer
              authConfig: { token: '{{token}}' }
              queryParams: ''
              headers: { Accept: application/json }
              variables: { orderId: '123', token: token-de-acesso }
              body: ''
      responses:
        '201': { description: Requisição salva }
  /api-integrations/requests/{requestId}:
    parameters: [{ $ref: '#/components/parameters/RequestId' }]
    patch:
      tags: [API Integrations]
      summary: Atualizar requisição salva
      operationId: updateApiRequest
      security: [{ cookieSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaveApiRequest' }
      responses:
        '200': { description: Requisição atualizada }
    delete:
      tags: [API Integrations]
      summary: Excluir requisição salva
      operationId: deleteApiRequest
      security: [{ cookieSession: [] }]
      responses:
        '200': { description: Requisição excluída }
  /api-integrations/requests/{requestId}/run:
    post:
      tags: [API Integrations]
      summary: Executar requisição salva
      operationId: runApiRequest
      security: [{ cookieSession: [] }]
      parameters: [{ $ref: '#/components/parameters/RequestId' }]
      responses:
        '201': { description: Resultado da execução }
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: connect.sid
  parameters:
...
    ChannelId: { in: path, name: channelId, required: true, schema: { type: integer, example: 1 } }
    RequestId: { in: path, name: requestId, required: true, schema: { type: integer, example: 1 } }
...
  schemas:
...
    CreateApiChannel:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        description: { type: string, maxLength: 240 }
    UpdateApiChannel:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        description: { type: string, maxLength: 240 }
    SaveApiRequest:
      type: object
      required: [name, method, url, authType]
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        description: { type: string, maxLength: 240 }
        method: { type: string, enum: [GET, POST, PUT, PATCH, DELETE] }
        url: { type: string, format: uri, maxLength: 2000 }
        authType: { type: string, enum: [none, bearer, basic, apiKey] }
        authConfig: { type: object, additionalProperties: { type: string } }
        queryParams: { type: string, maxLength: 4000 }
        headers: { type: object, additionalProperties: { type: string } }
        variables: { type: object, additionalProperties: { type: string } }
        body: { type: string, maxLength: 20000 }